+++ apps/web/src/common/cache.ts
@@ -1,1 +1,1 @@
-import {} from '@viness/core'
+import {} from '@viness/react'