+++ packages/viness-react/src/app.ts
@@ -1,13 +1,8 @@
-import { Container, createDecorator, IInstantiationService, ServiceIdentifier } from '@viness/di'
+import { SyncDescriptor } from '@viness/di'
 import { IVinessAppConfig, VinessAppConfig } from './app-config'
+import { container, IServiceContainer, IStoreContainer, ServiceContainer, StoreContainer } from './container'
 import { initI18n } from './i18n'
-import { VinessRouteObject, VinessRoute } from './route'
 import { IVinessRouter, VinessRouter } from './router'
-import { generateId } from './utils'
-
-const container = new Container()
-
-export const IVinessApp = createDecorator<VinessApp>('IVinessApp')
 
 /**
  * Manage all the instances
@@ -23,71 +18,28 @@ export class VinessApp {
         return container.get(IVinessRouter)
     }
 
-    /**
-     * add new service to app
-     */
-    addService<T, Services extends {}[]>(
-        id: ServiceIdentifier<T>,
-        service: new (...services: Services) => T,
-        isLazyInit: boolean = true
-    ) {
-        const supportType = isLazyInit ? 1 : 0
-        container.register(id, service, supportType)
+    get stores() {
+        return container.get(IStoreContainer)
     }
 
-    /**
-     * get service instance by id
-     *
-     * @param {ServiceIdentifier<T>} service identifier
-     * @returns {T} service instance
-     */
-    getService<T>(id: ServiceIdentifier<T>) {
-        return container.get(id)
-    }
-
-    /**
-     * add new route
-     *
-     * @param routeObj added route
-     * @param parentRouteId parent route of this added route
-     * @returns
-     */
-    addRoute(routeObj: VinessRouteObject, parentRouteId?: ServiceIdentifier<VinessRoute>) {
-        const vinessRouter = container.get(IVinessRouter)
-
-        const id = `VinessRoute_${routeObj.id || generateId()}`
-        const IRouteDecorator = createDecorator<VinessRoute>(id)
-
-        class Route extends VinessRoute {
-            constructor(@IInstantiationService instantiationService: IInstantiationService) {
-                super(routeObj, IRouteDecorator, instantiationService)
-            }
-        }
-
-        container.register(IRouteDecorator, Route)
-        vinessRouter.addRoute(IRouteDecorator, parentRouteId)
-
-        return IRouteDecorator
+    get services() {
+        return container.get(IServiceContainer)
     }
 }
 
 export function createVinessApp(config?: IVinessAppConfig) {
-    // Register the builtin features
-    // - aop
-    // - router
-    // - aoo configuration
-    container.register(IVinessApp, VinessApp)
+    // register the builtin features
+    container.register(IStoreContainer, StoreContainer)
+    container.register(IServiceContainer, ServiceContainer)
     container.register(IVinessRouter, VinessRouter)
-    container.register(IVinessAppConfig, VinessAppConfig)
-
-    container.get(IVinessAppConfig).setConfig(config)
+    container.register(IVinessAppConfig, new SyncDescriptor(VinessAppConfig, [config], true))
 
     // initialize features
     if (config?.i18n?.resources) {
         initI18n(config?.i18n)
     }
 
-    // Initialize the app
-    const app = container.get(IVinessApp)
+    // initialize the app
+    const app = new VinessApp()
     return app
 }