+++ packages/viness-react/src/http-client.ts
@@ -2,9 +2,9 @@ import axios, { AxiosRequestConfig, AxiosInstance } from 'axios'
 import { createDecorator } from './container'
 
 /**
- * Promise with cancel
+ * Promise can be canceled
  */
-export interface IServicePromise<T = any> extends Promise<T> {
+export interface ICancelablePromise<T = any> extends Promise<T> {
     /**
      * cancel the request
      */
@@ -13,7 +13,66 @@ export interface IServicePromise<T = any> extends Promise<T> {
 
 export const IHttpClient = createDecorator<IHttpClient>('IHttpClient')
 
-export interface IHttpClient {}
+export interface IHttpClient {
+    /**
+     * Get Request
+     *
+     * @param url
+     * @param config
+     */
+    get<P = any, T = any>(url: string, config?: AxiosRequestConfig): (params?: P) => ICancelablePromise<T>
+
+    /**
+     * Post
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    post<D = any, P = any, T = any>(
+        url: string,
+        config?: AxiosRequestConfig
+    ): (data?: D, params?: P) => ICancelablePromise<T>
+
+    /**
+     * Patch
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    patch<D = any, P = any, T = any>(
+        url: string,
+        config?: AxiosRequestConfig
+    ): (data?: D, params?: P) => ICancelablePromise<T>
+
+    /**
+     * Delete
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    delete<P = any, T = any>(url: string, config: AxiosRequestConfig): (params?: P) => ICancelablePromise<T>
+
+    /**
+     * Head
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    head<P = any, T = any>(url: string, config?: AxiosRequestConfig): (params?: P) => ICancelablePromise<T>
+
+    /**
+     * Options
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    options<P = any, T = any>(url: string, config?: AxiosRequestConfig): (params?: P) => ICancelablePromise<T>
+}
 
 /**
  * Http Client based in axios
@@ -31,22 +90,16 @@ export class HttpClient {
      * @param url
      * @param config
      */
-    protected get<P = any, T = any>(url: string, config?: AxiosRequestConfig) {
+    get<P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
         return (params?: P) => {
             const source = axios.CancelToken.source()
-
-            if (config) {
-                config.cancelToken = source.token
-            } else {
-                config = { cancelToken: source.token }
-            }
-
+            config.cancelToken = source.token
             config.params = params
 
             const promise = this.instance.get(url, config) as any
             promise.cancel = () => source.cancel()
 
-            return promise as IServicePromise<T>
+            return promise as ICancelablePromise<T>
         }
     }
 
@@ -57,19 +110,16 @@ export class HttpClient {
      * @param data
      * @param config
      */
-    protected post<P = any, T = any>(url: string, config?: AxiosRequestConfig) {
-        return (data?: P) => {
+    post<D = any, P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
+        return (data?: D, params?: P) => {
             const source = axios.CancelToken.source()
-            if (config) {
-                config.cancelToken = source.token
-            } else {
-                config = { cancelToken: source.token }
-            }
+            config.cancelToken = source.token
+            config.params = params
 
             const promise = this.instance.post(url, data, config) as any
             promise.cancel = () => source.cancel()
 
-            return promise as IServicePromise<T>
+            return promise as ICancelablePromise<T>
         }
     }
 
@@ -80,19 +130,16 @@ export class HttpClient {
      * @param data
      * @param config
      */
-    protected patch<P = any, T = any>(url: string, config?: AxiosRequestConfig) {
-        return (data?: P) => {
+    patch<D = any, P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
+        return (data?: D, params?: P) => {
             const source = axios.CancelToken.source()
-            if (config) {
-                config.cancelToken = source.token
-            } else {
-                config = { cancelToken: source.token }
-            }
+            config.cancelToken = source.token
+            config.params = params
 
             const promise = this.instance.patch(url, data, config) as any
             promise.cancel = () => source.cancel()
 
-            return promise as IServicePromise<T>
+            return promise as ICancelablePromise<T>
         }
     }
 
@@ -103,21 +150,16 @@ export class HttpClient {
      * @param data
      * @param config
      */
-    protected delete<P = any, T = any>(url: string, config?: AxiosRequestConfig) {
+    delete<P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
         return (params?: P) => {
             const source = axios.CancelToken.source()
-            if (config) {
-                config.cancelToken = source.token
-            } else {
-                config = { cancelToken: source.token }
-            }
-
+            config.cancelToken = source.token
             config.params = params
 
             const promise = this.instance.delete(url, config) as any
             promise.cancel = () => source.cancel()
 
-            return promise as IServicePromise<T>
+            return promise as ICancelablePromise<T>
         }
     }
 
@@ -128,21 +170,36 @@ export class HttpClient {
      * @param data
      * @param config
      */
-    protected head<P = any, T = any>(url: string, config?: AxiosRequestConfig) {
+    head<P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
         return (params?: P) => {
             const source = axios.CancelToken.source()
-            if (config) {
-                config.cancelToken = source.token
-            } else {
-                config = { cancelToken: source.token }
-            }
-
+            config.cancelToken = source.token
             config.params = params
 
             const promise = this.instance.head(url, config) as any
             promise.cancel = () => source.cancel()
 
-            return promise as IServicePromise<T>
+            return promise as ICancelablePromise<T>
+        }
+    }
+
+    /**
+     * Options
+     *
+     * @param url
+     * @param data
+     * @param config
+     */
+    options<P = any, T = any>(url: string, config: AxiosRequestConfig = {}) {
+        return (params?: P) => {
+            const source = axios.CancelToken.source()
+            config.cancelToken = source.token
+            config.params = params
+
+            const promise = this.instance.options(url, config) as any
+            promise.cancel = () => source.cancel()
+
+            return promise as ICancelablePromise<T>
         }
     }
 }