+++ packages/viness-react/src/index.tsx
@@ -1,8 +1,9 @@
 export * from './app'
-export * from './app-react'
-export * from './app-react-hooks'
-export * from './router-react'
-export * from './container'
+export * from './react/app-react'
+export * from './react/app-react-hooks'
+export * from './react/router-react'
 export * from './ui-store'
+export * from './http-client'
 
+export { createDecorator } from './container'
 export { useNavigate, useLocation, useParams, Outlet } from 'react-router-dom'