+++ packages/viness-react/src/react/app-react-context.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react'
-import { VinessApp } from './app'
+import { VinessApp } from '../app'
 
 const AppContext = createContext<VinessApp | null>(null)
 export const AppContextProvider = AppContext.Provider