+++ packages/viness-react/src/react/router-react.tsx
@@ -1,6 +1,5 @@
 import { RouterProvider } from 'react-router-dom'
 import { useAppContext } from './app-react-context'
-import { IVinessRouter, VinessRouter } from './router'
 
 /**
  * A React Component to create Routes
@@ -8,7 +7,7 @@ import { IVinessRouter, VinessRouter } from './router'
 export const AppRouter = () => {
     const app = useAppContext()
 
-    const vinessRouter = app?.getService(IVinessRouter) as VinessRouter
+    const vinessRouter = app?.router as any
     if (!vinessRouter) throw new Error('AppRouter is not wrapped b VinessApp')
 
     let router = vinessRouter?._getInnerRouter()