+++ packages/viness-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @viness/core
 
+## 0.2.0
+
+### Minor Changes
+
+- feat: add builtin HttpClient implement
+- add ServiceContainer to manage services
+- add storecontiner to manage stores
+
 ## 0.1.1
 
 ### Patch Changes